Error messages from OSTree daemon are now correct
Previously, the errors were incorrect due to a race condition
where the explicit state-changed signal was received before the
new value of the error code and message were received over dbus.
This is fixed by using the notify::state signal instead, which
will not come before the error message is updated.

[endlessm/eos-shell#1231]