Add script for upgrading ostree configuration
We currently have no way to change an existing installation's ostree
configuration. There are 2 reasons we'd want to do this:

  1. The machine should be moved to a new stable branch.
  2. The remote ostree URL or branch has changed in some way.

Both can be considered ostree upgrades. This script uses the new
set-origin subcommand of "ostree admin" to change both the repo
configuration and the origin refspec.

This tool is intended to run automatically. The current branch and url
are detected, and if they match a preconfigured list, they are updated
to new values.

[endlessm/eos-shell#4574]