Rename eos-updater to eos-autoupdater
The real updater is the daemon, and the current updater is only
responsible for driving the daemon automatically. To that end, rename
the program and services to eos-autoupdater.

To handle upgrades more safely, the systemd units provide aliases to
their old names. However, the configuration file and state directory
maintain their current names of /etc/eos-updater.conf and
/var/lib/eos-updater, respectively. Code would need to be written to
provide an upgrade path maintaing the system's state and configuration.

[endlessm/eos-shell#5087]