Rename ostree-daemon to eos-updater
Since the update daemon is not going upstream, rename it to an Endless
specific eos-updater. Now the update code consists of the update daemon
eos-updater and the autoupdating service eos-autoupdater.

In addition to renaming the programs, the DBus interface name has been
changed to com.endlessm.Updater. This will require changes in other
clients.

Internally, the namespace has changed from OTD to Eos with the main
object EosUpdater rather than OTDOSTree. The types, functions and macros
have also been changed accordingly. To be consistent,
ostree_daemon_set_state was renamed to eos_updater_set_state_changed
since it operates on the updater but would conflict with the generated
function.

[endlessm/eos-shell#5087]