build: Simplify gdbus-codegen rules
These were taken from a fancy Collabora template but then stripped a bit
for ease of use. This broke distcheck because I left the $(<F) but
removed the cd to the srcdir preceeding the command. Switch to standard
automake variables and simplify the rule a bit. Specify the generated
prefix directly and make sure the command only runs once.