Disallow downgrades
Report no updates are available if the "new" version doesn't
have a newer timestamp than the currently booted version.
This matches the behavior of "ostree admin upgrade", except
having an older version available isn't reported as an error.

If a downgrade is needed for testing,
"ostree admin upgrade --allow-downgrade" can be used.

[endlessm/eos-shell#2586]