Add support for product-specific OSTree branches
Use a product identifier to check against downloaded branch
configuration data whether this product has a product-specific OSTree
branch. If so, use this product-specific ref to check for updates,
otherwise use the contents from a fallback entry ([Default]).

We opted to download the data instead of relying on a local config file
because the later would imply a 2-step process for every branch
configuration change (one ostree upgrade to get an updated config file
and a second one to deploy the OSTree of the new branch).

In case we fail to obtain the branch configuration data from the remote
server, the update process is aborted (the server could be temporarily
down). The new deployment origin file will point to the ref that was
used for the upgrade.

[endlessm/eos-shell#6079]
https://phabricator.endlessm.com/T6079