Add a class for handling branch configuration data
This class stores the contents of the branch configuration data in two
forms - raw and parsed. Raw form is used for getting the checksum or
for validating its signature, as parsed form can not be used for that
(deserializing and serializing the data may not yield the exactly same
results). The parsed form is stored basically for convenience and
correctness (to make sure that the raw bytes we got are actually a key
file). Note that the correctness here does not mean if the branch file
contains the specific groups or fields - these checks are made during
the poll stage.

Other things stored are a sha512 sum of the raw contents, a GPG
signature of the raw contents and a download time.

The sha512 sum and the download time will be used only for the cases
when GPG signature is not provided. The sum is not stored on disk in
any form - it is computed from the raw contents. The download time is
stored as the modified time file attribute of the raw contents.

When signature is provided, the download time is NULL and the branch
file's timestamp should be taken from the branch file itself. There is
a helper function doing just that in eos-util.h -
eos_updater_get_timestamp_from_branch_file_keyfile.