Add a class for handling updater specific extensions
The extensions are the files stored under extensions/eos subdirectory
of the OSTree repository. This class is meant to be a single entity
for loading and saving them.

Currently it manages branch configuration data (with the help of
EosBranchFile), refs files with their signatures and finally summaries
as a fallback.

Refs are simple files that only contain the checksum of the latest
commit of a specific ref in a repository. Those files are signed, so
they originate from the main server.

As a fallback it also can store the summary file from the main server
as the means for retrieving the latest commit checksum in a ref. But
I'd really expect this to be scrapped when the refs file format is
agreed to.