Add a test-specific environment variable to the booted deployment getter
The EOS_UPDATER_TEST_UPDATER_DEPLOYMENT_FALLBACK environment variable,
if specified, tells the getter to use the first deployment from the
sysroot. This is because, during testing, there is no booted
deployment - sysroot returns such only when sysroot in actually in
/. Usually the first deployment returned from sysroot is the latest
one, which is a sane default for testing purposes. To make sure that
the booted deployment does not change during the lifetime of the
process (for example after applying an update), we remember the first
deployment and return it afterwards.

This also adds a convenience object for quit file handling, which can
be used to force some program to quit after the file was deleted.