Add test-specific environment variables to autoupdater
The EOS_UPDATER_TEST_UPDATER_QUIT_FILE environment variable, if
specified, tells the updater to quit if it is in a certain
state. Currently it quits if updater is in either none, ready, error
or update-applied state.

The EOS_UPDATER_TEST_UPDATER_USE_SESSION_BUS environment variable, if
specified, tells the updater to register its services on the session
bus, instead of on the system bus.