Add a test-specific environment variable to the update server
The EOS_UPDATER_TEST_UPDATE_SERVER_QUIT_FILE environment variable, if
specified, tells the server to quit the mainloop when it is idle and
the file was removed.

This also adds the quit file handling in the server. This change was
made to have an alternative to timeouts which are not really a great
thing when having a slow debugging session.