Rewrite poll for it to become a dispatcher
Currently Poll method reads configuration to get the order of the
download sources, then gets a list of source specific metadata
fetchers, calls them in the configured order, sends metrics and gets
the latest update info with the preference as specified by the source
order. From the update info, source addresses and extensions are
stored in updater data structure to be propagated to the next update
stage.