Add a poller fetching the update metadata from the main server
This is largely based on the already existing code. The updater gets
the source of the updates from OSTree repository configuration,
downloads the branch configuration data from one of the two locations
(the old one is from "/eos-branch", the new one is
"/extensions/eos/branch_file"). The rest of the steps are as described
in a commit adding common poll functions.