Add a poller fetching the update metadata from the main server
The updater gets the source of the updates from the avahi discoverer,
finds the latest branch configuration data based on txt-records and
prepares a list of URLs with the latest commit. The rest of the steps
are as described in a commit adding common poll functions.

Currently, there are two versions of the txt records:

- Version 1 contains OSTree path, time when this file was downloaded
  from the main server and the sha512 sum of the branch file. This
  version is mostly for the LAN servers that don't have their branch
  configuration data file signed. This is not really great, as
  trusting the unsigned data can end up installing malicious
  update. This is mostly meant for LAN server and I hope that it will
  be scrapped.

- Version 2 contains the OSTree path and the branch configuration data
  timestamp. The timestamp is normally taken from the branch
  configuration data itself (from the "UnixUTCTimestamp" field of the
  "main" group). This version is for the servers that have signed the
  branch configuration data files.

OSTree path is used to differentiate between different repositories
from the main server. It is mainly the path part of the OSTree
repository URL on the main server, without the leading slash (so it is
something like "ostree/eosfree-i386").