Build system changes
The major change here is the introduction of the two local libraries,
dbus and util. The dbus library contains the generated DBus service
and types, the util library contains the utilities.

autoupdater links only to the dbus library, eos-updater and
eos-update-server link to both.

This also adds the --disable-metrics configure flag, so the project can
be built also outside the Endless specific environment.

This also adds the --with-server-port configure flag to set the port
number on which the eos-update-server will listen to when it is
activated by systemd. The default is 12345.

This also adds all the new files to the build system.