Add avahi-specific classes and functions
This commits adds a service discoverer, a simple service struct and
several functions for generating the service file and for getting the
keys of the txt records.

The discoverer is looking for services of type
"_eos_updater._tcp". After the discovery is finished, it calls a given
callback with discovery results being a list of service structs.

The service struct holds the necessary data to connect to the service
(address and a port), name and domain of the service and some
additional information from the service served as txt records.