Add utilities for setting up the proper environment for testing updater
These functions and classes basically set up the main server with
subservers (EosTestServer and EosTestSubserver) and a client
(EosTestClient).

The EosTestServer only sets up by running a subserver setup and then
runs ostree trivial-httpd.

The EosTestSubserver is the actual ostree repository on the main
server. To set it up, we need an ostree path, devices, branch file
timestamp and a mapping of refs to commits. Some of the subserver
fields can be modified afterwards just to call the update function to
have some newer commits or branch files there.

The EosTestClient can be used to setup an initial client state by
taking a snapshot of the subserver. Then client can run either an
updater or an update server.

The tests also check some environment variables to run the updater or
the servers under gdb.

The EOS_CHECK_UPDATER_GDB_BASH_PATH environment variable, if
specified, tells the tests to generate the bash script that executes
the eos-updater in a valid environment at the given location.

The EOS_CHECK_UPDATE_SERVER_GDB_BASH_PATH_BASE environment variable,
if specified, tells the tests to generate the bash scripts that
execute the eos-updater in a valid environment. The value of the
environment variable is processed, so each sequence of 6 'X's is
replaced with a port number on which the server will listen to. It
also generates the ".deleteme" files which must be removed after the
server is started for the tests to continue.