Add service files for DBus daemon
This commit should not be necessary.

The first service file is for the eos updater, but we cannot use it as
we need to run it with a custom environment.

The second service file is for the gvfs daemon - without it I'm
getting test failures when deleting the temporary directory after the
test has otherwise passed.