Make ref file more secure
Instead of storing only a commit id in the ref file, store also a name
of the ref itself too. Since there is more than one field to store,
just make the ref file a key file with a "mapping" group that has
"commit" and "ref" fields.

This is to avoid scenario where malicious server has two valid and
signed ref files for refs A and B and puts the ref file for A under
the name of B.

With the ref name stored in the ref file we can actually check if the
ref name stored in the ref file and the ref file path are actually the
same.