Refactor download sources handling
In the beginning I thought that this configuration will be needed in
more than one stage, but in the end its use is contained only in
polling stage. This commit moves the download sources array from
EosUpdaterData to polling code.

It also adds a possibility to extend each type of the download source
with the source specific configuration which will be passed to the
fetcher as a GVariant of G_VARIANT_TYPE_DICT type. The keys and
variants stored in the dict are specific for each fetcher. The name of
the group for the source specific configuration is 'source "<name>"'.

Here we also simplify the "Order" configuration field. Instead of it
being a comma-separated list of source names, we make it a
semicolon-separated list of them. So we can just use the
g_key_file_{set,get}_string_list functions instead of inventing our
own functions.