Use libtool to execute binaries under gdb in make check
Currently, all the binaries we build do not link against any libtool
library we also build at the same time, so it doesn't fix any
immediate problem. But later we will introduce a library for preparing
an update on a volume and a small command line utility for that, so
running an it via libtool will be needed (unless we want to debug bash
running the libtool-generated script instead).