Split update-server reaping in tests into two parts
Currently we were removing the quit file and then waiting for the
update server to quit. Usually the timeout for checking the existence
of the quit file in the update server is around 5 seconds, so for many
instances of update server running it could take some time before all
of them are reaped.

To make it a bit faster we remove all the quit files first and then
wait for the update servers to quit. The timeouts in update servers
will likely happen around the same time in parallel.