Add a library for preparing updates on a volume
This adds some functions to prepare the update on a given volume based
on a given sysroot. From the sysroot, we take the repository and the
booted deployment, and from the booted deployment we take the refspec
and commit id. All those will be used to prepare the update.

There is also a function that takes the repo, refspec and commit id so
the update does not need to be from the booted deployment
specifically.

To prepare the usb update, we need to run the eos update server (or
similar) on the source repo. To avoid having another process for it,
most of the eos update server functionality is moved to the library
too, so we can run a server and do the pull in one process.

This makes the eos update server to link against the prepare lib.