Drop the last use of the generate_strv function in tests
In earlier revisions of the tests we had some hacks to avoid writing
long lists of local string variables. The generate_strv was a part of
those hacks. The hacks were replaced with a saner solution so it is
not needed anymore.

And the last use of it was simply unnecessary.