Disable all update steps on live sessions
eos-autoupdater.timer is already stopped and masked from the initramfs,
but this prevents any updates being fetched or installed.

The rationale here is that endless.live_boot sessions have a read-only
filesystem, with overlay+tmpfs mounted over directories which need to be
read-write. Running an ostree update on such a system is at best
wasteful and frustrating (it will be discarded as soon as you reboot, so
can never be used), and at worst disastrous (it fills /run, because
hardlinking files on overlay requires copying the file from the lower to
the upper filesystem, and your live system breaks until you reboot).

https://phabricator.endlessm.com/T13951