Refactor setting/clearing error state
I need to add a new representation for the current error state to the
API, and it will be much easier to do so correctly if setting and
clearing the current error is only done in one place.

Regarding the ORLY comment removed here, actually both parts are true:
eos_updater_set_state_changed() does indeed call
eos_updater_emit_state_changed(), *but* that function is a no-op when
the object has not yet been exported.

I also fixed a couple of GError leaks.