tests: Disable metrics submission when running tests
Disable eos-metrics from submitting any metrics from eos-updater from
the unit tests, using the new EOS_DISABLE_METRICS environment variable.

The metrics daemon doesn’t run in the test D-Bus session which the unit
tests use, which was causing eos-metrics to emit a warning when it
failed to submit some metrics. Since G_DEBUG=fatal-warnings was also set
in the unit tests, this caused the abortion of some of the subprocesses
of the unit test.

Even if the metrics daemon were running in the test D-Bus session, we
would not want it to submit spurious messages to the metrics server from
unit tests anyway.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T14894