tests: Use a per-test GPG home directory
Distribute the gpghome files and copy them into the test fixture’s
temporary directory for each test, as GPG might modify them (for
example, automatically converting from GPGv1 to GPGv2). The source
directory might be read-only (in which case this would fail), and writes
back to it make the tests less reproducible.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T14894