build: Add code coverage support
This uses the AX_CODE_COVERAGE macro from autoconf-archive, which is
already packaged for us. The Debian packaging will need to be updated to
add it as a build-dependency.

Our current code coverage is at 69% of lines, and 90% of functions.

To get code coverage reports, run:
    make -C $builddir/tests check-code-coverage
it will run `make check` and then generate a code coverage report from
it.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T14894