eos-repo-server: Add root path checks for serving files
Add some more thorough checks for the files served by the repo server,
to ensure that they are always within the server root directory, after
all paths have been canonicalised.

Note that for the moment, this does not resolve symlinks, so there is
still potentially a risk there if it’s possible for an untrusted user to
create symlinks within the server root directory (they could symlink to
/etc/shadow, for example). That requires further thought, especially to
be able to do it in a race-free manner.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T14894