eos-autoupdater: Use an unsigned variable for update_interval
The code explicitly checks for and rejects negative values of
update_interval as soon as it’s loaded from the config file, so make
that constraint explicit in the rest of the code by using guint rather
than gint.

Also take the opportunity to add some precondition assertions to
read_config_file() to make it more explicit that it expects non-NULL
arguments (or it will crash).

Also take the opportunity to rename update_interval to
update_interval_days everywhere to clarify its units.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T14894