eos-autoupdater: Add overflow checking to is_time_to_update() arithmetic
It’s unlikely to ever overflow, unless we reach the end of the world, or
the user inputs a massive value in the configuration file — but since
this is untrusted user data, we should tick all the boxes to be sure.

More realistically, I had to go through all the checking for this to
work out the safe limits of the time values here while validating the
code, so those calculations might as well now be encoded
programmatically.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T14894