eos-updater-data: Use a default cached branch file if none exists
If we are running the new version of eos-updater for the first time on a
system, /ostree/repo/extensions/eos/branch_file will not exist as a
cached copy of the branch_file (or its predecessor, eos-branch) from the
remote. As eos-branch seems to be empty on the remote anyway, start with
a default empty copy of it. It should be updated when the updater
successfully applies an update to the system.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T14894