eos-updater-avahi: Don’t ignore duplicate service names
The code previously ignored duplicate service names, which can happen
when two peers are both connected over IPv4 and IPv6: the same DNS-SD
name will be advertised over both interfaces. (It can also happen in a
variety of other ways.)

In any case, we should resolve and check *each* of these URIs, rather
than the first URI we find for a name, just in case some of them work
and others don’t. There might be problems with IPv6 on the local
network, for example. It’s unlikely, but it doesn’t hurt to return all
the URIs.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T14894