eos-updater-poll: Don’t use g_strconcat() for paths in URIs
Otherwise we can easily end up with double slashes (‘//’) separating
path components, which the path dispatcher in eos-update-server doesn’t
currently handle. Since it’s remotely possible we might want to handle
double slashes differently in URIs in the future, fix the code which
generates the request URIs. This is a closed system — we don’t need to
handle arbitrary user-provided URIs.

Generally, this means using functions which are separator-aware for
handling paths.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T14894