eos-updater-poll-common: Drop DOWNLOAD_N_SOURCES
This is not semantically a member of the enum, and its presence means
that -Wswitch-enum tries to make us handle it in every switch statement
for variables of this enum type. Avoid that by creating a DOWNLOAD_LAST
member instead, which has the same value as the last valid member of the
enum — -Wswitch-enum will ignore aliased members like this.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T14894