tests: Drop BASH_FUNC_ variables from test GDB script
They are not escaped correctly (due to containing newlines) and hence
break the generated script.

g_shell_quote() is not the right escape function to be using here, as
it’s designed for command line arguments. It’s not an escape function
for general SH quoting. This will do though.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T14894