eos-repo-server: Serve a minimal config file over HTTP to clients
The config file only needs to contain the repository mode and
repo_version, and none of the information about our remotes — clients
don’t need that, and it unnecessarily reveals username and password
information from the URIs (although those usernames and passwords are
typically the same everywhere anyway).

Instead, build a minimal config file and serve that. Double check that
the repository we’re reading from is in the right mode and version.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T14894