eos-updater: Move eos-updater.conf from /etc to /usr/share
If the contents of /etc/eos-updater.conf are the defaults, delete the
file on startup, and use the defaults from the new
/usr/share/eos-updater/eos-updater.conf instead. Otherwise, prefer the
configuration from /etc over /usr/share.

This is an attempt to eliminate the potential for future configuration
merge problems in /etc when we start to add more options to the
configuration file.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T14894