eos-autoupdater: Move configuration file to eos-autoupdater.conf
If the user manually edits the configuration file, they used to have to
save it as /etc/eos-updater.conf — but this file actually controls the
behaviour of eos-autoupdater. So move it to
/etc/eos-updater/eos-autoupdater.conf. If the new file isn’t found, but
the old one is, load it as a fallback. If neither are found, load the
defaults from /usr/share.

The new location in /etc is in the /etc/eos-updater directory so that we
can also rename /etc/eos-updater-daemon.conf to
/etc/eos-updater/eos-updater.conf without it colliding with the old name
for this file. This will be done in a following commit.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T14894