eos-prepare-usb-update: Use a mirroring pull
If we set the mirror option when pulling from the system repository into
the USB volume’s repository, we get the following things:

 1. If no refs to pull are specified, then all refs on the remote are
    pulled. That’s not the case here, as we specify a ref.
 2. Refs are pulled into the local namespace (refs/heads) rather than
    the remote namespace (refs/remotes), which means we can drop the
    code to convert them.
 3. The summary and signature files are copied from upstream if they’re
    available. This has little impact for us, as the EosExtensions are
    copied across regardless.
 4. Mirroring into an archive-z2 repo skips static deltas, as the exact
    objects are pulled, rather than being rebuilt with deltas. This will
    speed things up a bit.

As a result of #2, we can also drop the use of a local HTTP server to do
translations, which speeds up the repository preparation even more.
ostree has code itself to convert from objects bare to archive-z2 when
pulling into the USB repository.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T14894