eos-autoupdater: Don’t connect to NetworkManager in tests
Connecting to NetworkManager means connecting to the system bus, which
we might not have permission to do when running in unit tests. This
could cause construction of the NmClient to fail, which would take us
down with it.

Instead, don’t connect to NetworkManager when running the unit tests
(which use a mock session bus). Ideally we’d fix this by using a mock
system bus when running the tests, but GTestDBus doesn’t support that
(and adding support for it is not going to be accepted upstream: see bug
a fake bus instance (outside process) instead, but that’s work for
later.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T14894