tests: Drop D-Bus .service file for eos-updater
We do not want the updater to be bus-activated by the tests, as then it
will be launched with the wrong environment. Instead, we want
spawn_updater() in eos-test-utils.c to spawn it directly. Anything which
relies on the updater being on the bus before it’s spawned explicitly
will now fail, rather than D-Bus activating it in the wrong environment.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T14894