util: Add ostree helper functions to get advertisable commits
These are the commits which we will advertise over the local network as
being available to download from this peer. Currently,
src/eos-updater-avahi.c advertises the commit which has just been
applied as an update; that code will be removed shortly and replaced
with some which uses eos_sysroot_get_advertisable_commit().

eos_sysroot_get_advertisable_commit() returns the most recent commit
from all the sysroot’s deployments of the same OS as is currently
booted. This has been chosen rather than the most recent commit out of
all available refs, because we can more reliably guarantee that the
objects referenced by a deployment will not disappear as part of a
cleanup operation — since they’re either in use, or are about to be
booted into (if we’ve just applied an update but haven’t yet rebooted).

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T15544