tests: Add basic integration test for eos-updater-avahi
This is an installed-only integration test which pokes the actual
configuration files on a deployed system to see if the Avahi service
advertisement changes accordingly.

Only a single test case is included in the test at the moment — this
commit is mostly about adding the test framework and build/packaging
integration.

This adds a MIT-licensed file from pycotap: taptestrunner.py. Its
inclusion is compatible with its MIT license. We could alternatively not
copy it into our code base; that would require packaging pycotap and
adding a dependency on it. This seems like a bit of a waste of time for
~180 lines of Python.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T15544