eos-updater-poll: Consistently error out of Poll() if not using OSTree
If on a dev-converted system, consistently error out of Poll() with a
specific error code which is advertised over D-Bus. Previously, it was
left to individual metadata fetchers to error out when they got to a
point where not using OSTree was going to cause problems, and they used
non-standard error codes which were not mapped to D-Bus error names,
meaning they weren’t especially useful over the bus.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T155644