eos-update-server: Use more descriptive exit statuses
Instead of using exit status 1 to describe all errors, use a variety of
well-defined statuses, so that integration tests (to be added in a
following commit) can easily distinguish between failure modes.

Note that the enum introduced in this commit includes some error codes
which are not yet used in the code. They will be used in following
commits.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T15544