eos-update-server: Read configuration file to see if enabled
Read /etc/eos-updater/eos-update-server.conf to check whether
advertising updates over the local network is enabled. If it’s not
enabled on startup, exit from the server.

Currently the configuration file is not monitored for changes throughout
the life of the server process, but might be in future.

This shares its configuration file with eos-updater-avahi.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T15544