data: Add Also= lines to link .service to .timer and .socket files
This means that `systemctl [enable|disable]` on any of these units will
enable/disable the linked ones. Although note that any unit lacking a
WantedBy= key in [Install] is a static service, and cannot be enabled or
disabled anyway. It’s good to include the information for semantic
purposes though.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T15544