eos-updater-ctl: Fix monitoring after initial state change
As we now set up the proxy when we’ve found the updater’s name on the
bus, it all happens in a callback function. When the callback function
finishes, the proxy goes out of scope and is destroyed — so we get no
signals for property changes.

Fix that by keeping the proxy around in a global variable until the name
is lost.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T15544