tests: Improve check for skipping ‘unreadable file’ test
Previously this was testing for whether the effective UID was root, on
the assumption that root was the only user able to bypass DAC checks on
file permissions. However, on Jenkins it seems this is not the case —
there might be some UID remapping going on, or the jenkins user might
have CAP_DAC_OVERRIDE. In any case, the check can be improved by testing
to see if the user can read the ‘unreadable file’. If they can, skip the
test.

The test is run frequently enough locally by anyone developing
eos-updater that it will still be run enough to catch problems with the
code.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T15545