tests: Fix an infinite loop hang waiting for eos-update-server
If the eos-update-server child process crashes or exits early while
run_update_server() is waiting for it to output a port file, the test
will sit in a `while (TRUE) sleep (1);` loop forever. Avoid that by
adding a 10s timeout, which should be plenty of time for the child
process to start and create the port file.

If the user is running eos-update-server under gdb, don’t impose a limit
on the wait.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T15544