eos-repo-server: Change some debug messages to warnings
Change some debug messages to warnings, as otherwise the problem is only
indicated using HTTP status codes sent to the client. The warnings will
end up in the journal, for later analysis.

Also change a g_message() to a g_debug(), otherwise it floods the
journal with irrelevant checksums for each successful file transfer.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T15544