eos-autoupdater: Add a randomised delay in starting the autoupdater
Add a randomised delay of between 0 and 30 minutes to each time the
autoupdater is started, using the systemd RandomizedDelaySec= key (see
`man systemd.timer`). The delay is chosen uniformly from that interval,
so across a classroom of 30 machines all polling the teacher’s machine
for updates, we can expect one poll request every minute, on average,
with this change applied.

If an update is available, this means the teacher’s machine will have
to, on average, serve it to one additional student machine every minute.
If it takes 5 minutes to transfer the update objects across the LAN (a
conservative estimate), the teacher’s machine will have to, on average,
handle 5 concurrent download clients.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T15870