eos-update-server: Deprioritise eos-update-server for CPU and I/O
Set the niceness and I/O scheduling class so that eos-update-server gets
idle resources only, so that it doesn’t impact the interactivity of a
machine if several peers are downloading updates from it over the
network. Currently, eos-update-server compresses objects on the fly as
it sends them to peers, which imposes a heavy CPU load. This is an
attempt to mitigate that.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T15870