eos-updater-avahi: Make the .service file a oneshot service
This should ensure that any services which depend on it only run once
it’s updated the Avahi service file — rather than as soon as it’s been
exec()ed, which might be before it gets a chance to update the .service
file. This fixes a potential race condition in sequencing unit startup,
although the race likely does not occur in practice because nothing
depends on eos-updater-avahi.service. It might make the integration
tests more reliable.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T15870