eos-updater-ctl: Block on the requested command completing
Change the behaviour of `eos-updater-ctl $command` to block on the
command completing (successfully or not), rather than returning as soon
as the initial D-Bus call to eos-updater returns. This should make
eos-updater-ctl more useful from the command line for scripting.

A follow-up commit will restore the original behaviour under a new
--no-block argument.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T15870