eos-updater: Fix implementation of D-Bus PollVolume method
While the code to implement this method exists and is compiled, it was
never hooked up to the signal callback from GDBus, so calls to it would
result in org.freedesktop.DBus.Error.UnknownMethod. This might have been
a rebase/merge problem.

It affected the implementation of `eos-autoupdater --from-volume=` and
not much else — nothing else calls PollVolume (yet).

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T15870