eos-updater-poll-lan: Fix crash polling a LAN source on the same ref
If eos-updater polled for updates from another computer on the local
network which has the same (or an older) OSTree ref deployed as the
current computer, eos-updater would crash.

This never actually caused a problem, because eos-updater not running is
essentially equivalent to it running and being in the Ready state, which
is the state it would have switched back to after finding no newer
OSTree commits to fetch from the other computer. It would have caused
problems on networks where there’s more than one computer advertising
updates; but that’s not a configuration we support yet.

In any case, crashing is bad. Fix it. The confusion resulted from
is_checksum_an_update() returning TRUE while explicitly not returning
something in one of its out parameters, and the caller not handling this
situation correctly.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T15908