eos-repo-server: Set compression level to 2 rather than 9
When sending OSTree blobs over the network, we compress them on the fly
to convert from bare format into archive-z2 format. Previously, we were
using the default compression level which, for our version of OSTree, is
9 — the highest zlib compression level.

This results in a lot of CPU usage, which is not good when multiple
clients are simultaneously downloading from this computer. It makes
eos-update-server consume all of the available CPU.

Instead of compression level 9, use compression level 2. This requires a
third of the CPU time, giving compression which is about 90% as good as
level 9. It’s worth noting that upstream OSTree has since switched to
level 6 by default.

This depends on a new API in OSTree,
ostree_raw_file_to_archive_z2_stream_with_options(), which will be
available in the 2017.3 release. We’ve backported it to our OSTree
package, which is at 2016.14. Hence the version requirement in
eos-updater has not been bumped even though it technically should be.

https://phabricator.endlessm.com/T15909