eos-autoupdater: Add a randomised delay to the IntervalDays=
As discussed in T15975, to avoid computers falling into the same pattern
of updating in sync every 2 weeks, and to avoid a large deployment of
computers all attempting to update at the same time when initially
turned on; add a random delay to the update interval calculation.

By default this delay is 0 days (i.e. not applied), so that computers
which are directly connected to the internet are updated as soon as they
are first turned on. The idea is that large customised deployments of
Endless OS could set RandomizedDelayDays=7 (for example) to apply a
delay uniformly distributed between 0 and 7 days (inclusive).

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T15975