eos-updater-avahi: Add Wants= dependency from multi-user.target
Previously, the eos-updater-avahi.service unit would only be triggered
when any of the paths watched by eos-updater-avahi.path were changed.
This isn’t actually what we want — we also want it to be triggered on
boot when those paths aren’t changed, just to sync the state of the
Avahi service file.

Add a WantedBy=multi-user.target to achieve that; and correctly add the
Wants=avahi-daemon.service relation to ensure this does pull the Avahi
daemon into the boot too.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T16000