eos-updater-avahi: Wait for avahi-daemon to start when polling the LAN
When constructing an AvahiClient, construction can fail if the
avahi-daemon is not already running. Pass the AVAHI_CLIENT_NO_FAIL flag
to ensure that it waits for the daemon to start (and handles situations
where the avahi-daemon is restarted while eos-updater is still running).

See http://www.avahi.org/doxygen/html/#daemon_dies.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T16095