eos-updater-avahi: Don’t stop polling the LAN when done with the cache
When browsing for DNS-SD services on the local network, avahi-daemon
will first return all cached services to the Avahi client (eos-updater).
It will then emit the AVAHI_BROWSER_CACHE_EXHAUSTED event, before moving
on to actively query the network for more up to date services.

eos-updater was aborting the browse operation when the CACHE_EXHAUSTED
event was received — so an active query of services on the LAN would
never happen. This means that eos-updater could never see updater
services on the network which were not actively advertised (i.e.
started) while eos-updater was running. So in the situation where a
server has been running for a long time and the client (or avahi-daemon
on the client) restarts, the client will no longer be able to see the
server until it next periodically advertises itself on the network.

Fix that by ignoring the CACHE_EXHAUSTED event, and only stopping the
browse operation on receiving AVAHI_BROWSER_ALL_FOR_NOW.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T16095