eos-update-server: Change default timeout from 5s to 200s
Make the server exit timeout longer than the TCP connection timeout, so
the server doesn’t exit while a client on a lossy connection is
potentially resending a request to it which it never received due to
packet loss. This ensures that the client correctly reports such errors
as a timeout, rather than an unexpectedly terminated connection. Or,
even better, the client’s resend might succeed and the request would be
handled successfully.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T16387