eos-repo-server: Use request-read rather than request-started signals
The documentation for request-started is unclear, but it turns out to be
emitted unconditionally when SoupServer starts listening on a socket (or
unconditionally after it finishes processing a request on a keepalive
socket). i.e. It’s emitted in anticipation of the SoupServer eventually
reading another message.

For the purposes of counting the number of requests the repo server
actually processes, we want request-read, which is emitted after the
body of a request has been read from the socket.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T16387