eos-repo-server: Make it derive from GObject rather than SoupServer
Instead, store the SoupServer as a private pointer, and use it that way.
This means we will soon be able to have multiple EosUpdaterRepoServer
instances using a single SoupServer, each handling different paths.

Rename a lot of variables from `server` to `self` to be consistent with
the rest of the code.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T16551