libeos-update-server: Add EusServer to encapsulate multiple EusRepos
At the moment, this is only used by eos-update-server to serve a single
repo, but it restructures the code so that serving multiple repos should
be easy in future.

The pending-requests and last-request-time properties are moved from
EusRepo to EusServer, since they’re per-SoupServer.

This adds various new bits of API documentation, but no new tests yet.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T16551