eos-updater: Handle G_IO_ERROR_FAILED when running in FBE test mode
When test mode is enabled in the first boot environment, overlay file
systems are mounted on top of everything (including /ostree), which
breaks OSTree. (This is expected.)

This causes libostree to return an error from ostree_sysroot_load(),
which we were not correctly handling, and hence were passing through the
wrong error domain over eos-updater’s D-Bus interface. This caused an
error message to appear in the UI when it wasn’t supposed to.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T16574