Pass refspec to fetch_latest_commit()
Instead of taking a remote and ref name, take a full refspec string.

This avoids having to store the remote and ref in the 3 callsites,
and also will keep the interface clean and symmetrical after the next
commit where this function will be extended to return a (possibly
changed) refspec.