Follow ref redirects in commit metadata
When a commit has the ostree.endoflife-rebase metadata set,
we treat it as a branch redirect and proceed to update from the redirect
target branch instead.

This will be used for migrating nexthw users to the mainline releases
at the appropriate time.

This implements the same logic that went upstream in
OstreeSysrootUpgrader.

https://phabricator.endlessm.com/T17339