eos-updater-apply: Workaround failures during sysroot cleanup
The sysroot cleanup happens after the successful deployment. The
failures are likely the result of the concurrent prunes that happened
earlier.

So tell ostree to skip cleanup and perform it ourselves later, so we
can simply log an error, but still report a success as the "apply"
command is mostly about creating a new deployment, not doing the
cleanup.