tests: Generate more files and directories
This is for the cleanup workaround test, where we want to remove some
.dirtree objects which ostree will want to prune, but so far we didn't
have any real directory structure and thus not a single dirtree object
in the repository.

So this commit will generate directories and files fitting this
scheme for a commit X:

/for-all-commits/commit(0…X).dir/{a,b,c}/{x,y,z}.X

That should generate a plenty of objects to play with.