tests: Add a cleanup workaround test
This tries to reproduce the issue with a corrupted repository where
some files that will be pruned are missing, so the cleanup happening
after the update has been deployed fails, thus making the user think
that the whole operation failed.

It is rather tricky to reproduce and I am still not sure this scenario
succeeds, because the warning does not happen anymore. This may be
because either I did it wrong in the test or ostree has changed its
code to not bail out when some metadata object is missing.