tests: Generate delta files on the server
When updating the server, we will generate the delta files between the
old and new commits. For it to be useful, we generate a big file (10MB
large), where we change just one and the same byte on each commit.

Since the broken delta test involves removing the generated big file,
we need to know its location in the deployment root, so we can learn
about the big file's checksum.