euu/config: Expand config API to support per-key fallbacks
The previous incarnation of the configuration file handling supported
per-file fallbacks: it would load all the configuration keys from one
file or, if that didn’t exist, fall back to the next one. However, it
did not support falling back to the next file if a specific key did not
exist.

That poses problems for newly introduced keys in configuration files,
where the user might have copied the old configuration file to make
customisations — the old file doesn’t have the new key, which leaves the
code unhappy (as it expects to always be able to load a key it queries
for, since they are all expected to be listed in the default
configuration file, which *has* to be installed and read-only).

Fix that by reworking the configuration API to load multiple files in
parallel and handle per-key fallbacks between them. The new API mostly
mirrors the GKeyFile API, although with a more modern type system
modelled on GVariant, and some inbuilt validation support.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T18308